DMM-011 linked db with layout [HM,MG]

+++ app/src/main/java/com/team13/dealmymeal/core/Plan.kt
@@ -10,7 +10,7 @@ import com.team13.dealmymeal.data.Meal
 class Plan(
     @ColumnInfo(name = "period") var period: Int,
     @ColumnInfo(name = "created_time", defaultValue = "CURRENT_TIMESTAMP") var createdTime: String,
-    @ColumnInfo(name = "meals") var meals: List<Int>?
+    @ColumnInfo(name = "meals") var meals: List<Long>?
 ) {
     @PrimaryKey(autoGenerate = true)
     var id: Long = 0L
